refactor: simplify store
